+++ lib/constants/webrtc_configurations.dart
@@ -1,22 +1,20 @@
+import 'package:flutter/foundation.dart';
+
 import 'package:waterbus_sdk/flutter_waterbus_sdk.dart';
 
 class WebRTCConfigurations {
   static const Map<String, dynamic> configurationWebRTC = {
     'iceServers': [
       {
-        "urls": "stun:turn.waterbus.tech:3478",
+        "urls": "stun:stun.cloudflare.com:3478",
       },
-      {
-        "urls": "turn:turn.waterbus.tech:3478?transport=udp",
-        "username": "waterbus",
-        "credential": "waterbus",
-      }
     ],
     'sdpSemantics': 'unified-plan',
     'iceCandidatePoolSize': 20,
     "audioJitterBufferMaxPackets": 50,
     'bundlePolicy': 'max-bundle',
     "rtcpMuxPolicy": "require",
+    'encodedInsertableStreams': kIsWeb,
   };
 
   static const Map<String, dynamic> offerPublisherSdpConstraints = {
@@ -41,14 +39,14 @@ class WebRTCConfigurations {
       maxBitrate: 2500000,
       minBitrate: 1500000,
       maxFramerate: 30,
-      scalabilityMode: 'L1T3',
+      // scalabilityMode: 'L1T3',
     ),
     RTCRtpEncoding(
       rid: 'h',
       maxBitrate: 900000,
       minBitrate: 500000,
       maxFramerate: 24,
-      scalabilityMode: 'L1T2',
+      // scalabilityMode: 'L1T2',
       scaleResolutionDownBy: 2.0,
     ),
     RTCRtpEncoding(
@@ -56,7 +54,7 @@ class WebRTCConfigurations {
       maxBitrate: 300000,
       minBitrate: 100000,
       maxFramerate: 15,
-      scalabilityMode: 'L1T1',
+      // scalabilityMode: 'L1T1',
       scaleResolutionDownBy: 4.0,
     ),
   ];