+++ lib/stats/webrtc_audio_stats.dart
@@ -1,10 +1,11 @@
 import 'dart:async';
 
+import 'package:flutter_webrtc_plus/flutter_webrtc_plus.dart';
 import 'package:injectable/injectable.dart';
 
-import 'package:waterbus_sdk/flutter_waterbus_sdk.dart';
 import 'package:waterbus_sdk/types/enums/audio_level.dart';
 import 'package:waterbus_sdk/types/models/audio_stats_params.dart';
+import 'package:waterbus_sdk/types/models/stats.dart';
 import 'package:waterbus_sdk/utils/extensions/duration_extensions.dart';
 
 @singleton
@@ -17,13 +18,34 @@ class WebRTCAudioStats {
     _sender = param;
   }
 
-  void addReceiver(AudioStatsParams param) {
-    _receivers.add(param);
+  void addReceiver({
+    required String ownerId,
+    required RTCRtpReceiver receiver,
+    required Function(AudioLevel) callback,
+  }) {
+    final int index = _receivers.indexWhere(
+      (params) => params.ownerId == ownerId,
+    );
+
+    if (index < 0) {
+      _receivers.add(
+        AudioStatsParams(
+          ownerId: ownerId,
+          callBack: callback,
+          receivers: [receiver],
+        ),
+      );
+    } else {
+      _receivers[index] = _receivers[index].copyWith(
+        receivers: [receiver],
+        callBack: callback,
+      );
+    }
   }
 
-  void removeReceiver(String peerConnectionId) {
+  void removeReceiver(String ownerId) {
     final int index = _receivers.indexWhere(
-      (params) => params.peerConnection.peerConnectionId == peerConnectionId,
+      (params) => params.ownerId == ownerId,
     );
 
     if (index < 0) return;
@@ -32,7 +54,7 @@ class WebRTCAudioStats {
   }
 
   void initialize() {
-    _timer ??= Timer.periodic(500.milliseconds, (timer) {
+    _timer ??= Timer.periodic(1.seconds, (timer) {
       if (_sender != null) _monitorAudio(params: _sender!);
 
       for (final params in _receivers) {
@@ -56,8 +78,8 @@ class WebRTCAudioStats {
     final List<StatsReport> stats = [];
 
     if (type == 'media-source') {
-      final List<StatsReport> senderStats =
-          await params.peerConnection.getStats();
+      if (params.pc == null) return;
+      final List<StatsReport> senderStats = await params.pc!.getStats();
       stats.addAll(senderStats);
     } else {
       final List<RTCRtpReceiver> rtpReceivers = params.receivers;