+++ lib/types/enums/rtc_track_kind.dart
@@ -1,0 +1,7 @@
+enum RtcTrackKind {
+  audio('audio'),
+  video('video');
+
+  final String kind;
+  const RtcTrackKind(this.kind);
+}