+++ lib/types/enums/track_type.dart
@@ -1,0 +1,14 @@
+enum TrackType {
+  webcam,
+  screen;
+}
+
+extension TrackTypeX on int? {
+  TrackType get trackType {
+    return switch (this) {
+      0 => TrackType.webcam,
+      1 => TrackType.screen,
+      _ => TrackType.webcam
+    };
+  }
+}